move p2p mock into its own file
